+++ src/commonMain/kotlin/GameModule.kt
@@ -8,7 +8,7 @@ import kotlin.reflect.*
 object GameModule : Module() {
 
     override val mainScene: KClass<out Scene> = TestScene::class
-    override val size: SizeInt = SizeInt(1280,720)
+    override val size: SizeInt = SizeInt(1920,1080)
     //override val bgcolor: RGBA = Colors.WHITE
 
     override suspend fun AsyncInjector.configure() {
